Unify free-function facades into nested Cv2.<Sub> classes (#1921)
Replace the standalone sub-namespace facade classes (CvDnn, CvAruco,
CvXImgProc, CvXPhoto, CvOptFlow, CvText, CvDetail) with nested static
partial classes under Cv2 (Cv2.Dnn, Cv2.Aruco, ...), mirroring the C++
sub-namespaces (cv::dnn::, cv::aruco::, ...) and the Python cv2.* submodules.

- Wrapper types stay in their OpenCvSharp.<Sub> namespaces; only the
  free-function facades move. Each facade file is renamed to Cv2.<Sub>.cs
  in its module folder.
- CvShape (no cv::shape sub-namespace) folds directly into Cv2.
- CvExtensions (OpenCvSharp.Extensions extension methods) is left as-is.
- Clean break: the old CvXxx classes are removed (no obsolete forwarders),
  since OpenCvSharp5 is a new package family with no installed base.
- Cv2.XPhoto.Inpaint signature now fully-qualifies OpenCvSharp.XPhoto.InpaintTypes
  to avoid binding to the same-named OpenCvSharp.InpaintTypes.
- Update all test call sites, the 4->5 migration guide, dnn coverage notes,
  and document the "Cv2 = cv::, Cv2.<Sub> = cv::<sub>::" rule in
  copilot-instructions.
