+++ docs/migration-4-to-5.md
@@ -42,6 +42,37 @@ core **`OpenCvSharp`** package. Its namespace is unchanged
 (`OpenCvSharp.Extensions`), so existing `using OpenCvSharp.Extensions;` keeps
 working — but you no longer need a separate extensions package for it.
 
+### Sub-namespace free functions: `CvXxx` → `Cv2.Xxx`
+
+The standalone free-function facade classes (`CvDnn`, `CvAruco`, …) have been
+replaced by **nested static classes under `Cv2`**, so the managed API mirrors the
+C++ sub-namespaces and the Python `cv2` submodules (`cv2.dnn.*`, `cv2.aruco.*`, …):
+
+| Before (OpenCvSharp4) | After (OpenCvSharp5) | C++ namespace |
+|---|---|---|
+| `CvDnn.ReadNetFromOnnx` | `Cv2.Dnn.ReadNetFromOnnx` | `cv::dnn` |
+| `CvAruco.DetectMarkers` | `Cv2.Aruco.DetectMarkers` | `cv::aruco` |
+| `CvXImgProc.NiblackThreshold` | `Cv2.XImgProc.NiblackThreshold` | `cv::ximgproc` |
+| `CvXPhoto.Inpaint` | `Cv2.XPhoto.Inpaint` | `cv::xphoto` |
+| `CvOptFlow.CalcOpticalFlowSF` | `Cv2.OptFlow.CalcOpticalFlowSF` | `cv::optflow` |
+| `CvText.DetectTextSWT` | `Cv2.Text.DetectTextSWT` | `cv::text` |
+| `CvDetail.ComputeImageFeatures` | `Cv2.Detail.ComputeImageFeatures` | `cv::detail` |
+| `CvShape.CreateThinPlateSplineShapeTransformer` | `Cv2.CreateThinPlateSplineShapeTransformer` | `cv::` (no sub-namespace) |
+
+The wrapper **types** (`Net`, `ArucoDetector`, the `ximgproc` filter classes, …)
+are unchanged and still live in their `OpenCvSharp.<Sub>` namespaces, so a typical
+call site only changes the static facade prefix:
+
+```csharp
+using OpenCvSharp.Dnn;                       // Net / Model types (unchanged)
+var net = Cv2.Dnn.ReadNetFromOnnx("m.onnx"); // was: CvDnn.ReadNetFromOnnx(...)
+```
+
+`CvShape` had no matching `cv::shape` sub-namespace (its members are `Create…`
+factories for types living directly in `cv::`), so it folds straight into `Cv2`.
+`CvExtensions` (the OpenCvSharp-specific `Mat` extension methods) is **not** part
+of this change and keeps its name and `OpenCvSharp.Extensions` namespace.
+
 ## 3. Removed APIs
 
 These wrap OpenCV functions that were removed in OpenCV 5 (no replacement), so
@@ -49,8 +80,8 @@ the managed methods are gone too:
 
 | Removed | Notes / replacement |
 |---|---|
-| `OpenCvSharp.Extensions.Binarizer` (Niblack/Sauvola/Bernsen/Nick/…) | Use `CvXImgProc.NiblackThreshold` (Niblack / Sauvola / Wolf / Nick) |
-| `CvDnn.ReadNetFromDarknet`, `CvDnn.ReadNetFromCaffe`, `CvDnn.ReadNetFromTorch`, `CvDnn.ReadTorchBlob`, `CvDnn.ShrinkCaffeModel` (and the `Net.*` equivalents) | Darknet/Caffe/Torch parsers removed — export to ONNX and use `CvDnn.ReadNetFromONNX` |
+| `OpenCvSharp.Extensions.Binarizer` (Niblack/Sauvola/Bernsen/Nick/…) | Use `Cv2.XImgProc.NiblackThreshold` (Niblack / Sauvola / Wolf / Nick) |
+| `CvDnn.ReadNetFromDarknet`, `CvDnn.ReadNetFromCaffe`, `CvDnn.ReadNetFromTorch`, `CvDnn.ReadTorchBlob`, `CvDnn.ShrinkCaffeModel` (and the `Net.*` equivalents) | Darknet/Caffe/Torch parsers removed — export to ONNX and use `Cv2.Dnn.ReadNetFromONNX` |
 | `Net.SetHalideScheduler` | Halide backend removed |
 | `Cv2.ConvertFp16` | Use `Mat.ConvertTo` to/from `MatType.CV_16F` |
 | `Cv2.LogPolar`, `Cv2.LinearPolar` | Use `Cv2.WarpPolar` (`WarpPolarMode.Linear` / `Log`) |
