+++ test/OpenCvSharp.Tests/dnn/DnnTest.cs
@@ -17,7 +17,7 @@ public class DnnTest : TestBase
         float scoreThreshold = 0.5f;
         float nmsThreshold = 0.4f;
 
-        CvDnn.NMSBoxes(bboxes, scores, scoreThreshold, nmsThreshold, out var indices);
+        Cv2.Dnn.NMSBoxes(bboxes, scores, scoreThreshold, nmsThreshold, out var indices);
 
         Assert.Equal(2, indices.Length);
         Assert.Equal(0, indices[0]);
