+++ test/OpenCvSharp.Tests/dnn/TensorflowTest.cs
@@ -17,11 +17,11 @@ public class TensorflowTest : TestBase
     {
         using var imgOf9 = LoadImage(Path.Combine("Dnn","MNIST_9.png"), ImreadModes.Grayscale);
             
-        var img9DataBlob = CvDnn.BlobFromImage(imgOf9, 1f / 255.0f);
+        var img9DataBlob = Cv2.Dnn.BlobFromImage(imgOf9, 1f / 255.0f);
         var modelPath = Path.Combine("_data", "model", "MNISTTest_tensorflow.pb");
         var res = -1;
 
-        using (var tfGraph = CvDnn.ReadNetFromTensorflow(modelPath))
+        using (var tfGraph = Cv2.Dnn.ReadNetFromTensorflow(modelPath))
         {
             Assert.NotNull(tfGraph);
             tfGraph!.SetInput(img9DataBlob);
@@ -39,13 +39,13 @@ public class TensorflowTest : TestBase
     {
         using var imgOf5 = LoadImage(Path.Combine("Dnn", "MNIST_5.png"), ImreadModes.Grayscale);
 
-        var img5DataBlob = CvDnn.BlobFromImage(imgOf5, 1f / 255.0f);
+        var img5DataBlob = Cv2.Dnn.BlobFromImage(imgOf5, 1f / 255.0f);
         var modelPath = Path.Combine("_data", "model", "MNISTTest_tensorflow.pb");
         var res = -1;
 
         using (var stream = new FileStream(modelPath, FileMode.Open))
         {
-            using (var tfGraph = CvDnn.ReadNetFromTensorflow(stream))
+            using (var tfGraph = Cv2.Dnn.ReadNetFromTensorflow(stream))
             {
                 Assert.NotNull(tfGraph);
                 tfGraph!.SetInput(img5DataBlob);
