+++ test/OpenCvSharp.Tests/stitching/Cv2DetailTest.cs
@@ -5,7 +5,7 @@ using Xunit;
 
 namespace OpenCvSharp.Tests.Stitching;
 
-public class CvDetailTest: TestBase
+public class Cv2DetailTest: TestBase
 {
     // Platform check for conditional test execution
     public static bool IsWindowsOrLinux => RuntimeInformation.IsOSPlatform(OSPlatform.Windows) || RuntimeInformation.IsOSPlatform(OSPlatform.Linux);
@@ -17,7 +17,7 @@ public class CvDetailTest: TestBase
         using var featuresFinder = AKAZE.Create();
         using var image = LoadImage("abbey_road.jpg", ImreadModes.Grayscale);
 
-        using var features = CvDetail.ComputeImageFeatures(featuresFinder, image);
+        using var features = Cv2.Detail.ComputeImageFeatures(featuresFinder, image);
         Assert.NotNull(features);
         //Assert.NotEqual(0, features.ImgIdx);
         Assert.Equal(image.Size(), features.ImgSize);
@@ -33,8 +33,8 @@ public class CvDetailTest: TestBase
         using var image1 = LoadImage("tsukuba_left.png", ImreadModes.Grayscale);
         using var image2 = LoadImage("tsukuba_right.png", ImreadModes.Grayscale);
 
-        using var features1 = CvDetail.ComputeImageFeatures(featuresFinder, image1);
-        using var features2 = CvDetail.ComputeImageFeatures(featuresFinder, image2);
+        using var features1 = Cv2.Detail.ComputeImageFeatures(featuresFinder, image1);
+        using var features2 = Cv2.Detail.ComputeImageFeatures(featuresFinder, image2);
 
         using var matcher = new BestOf2NearestMatcher();
         using var matchesInfo = matcher.Apply(features1, features2);
@@ -51,8 +51,8 @@ public class CvDetailTest: TestBase
         using var image1 = LoadImage("tsukuba_left.png", ImreadModes.Grayscale);
         using var image2 = LoadImage("tsukuba_right.png", ImreadModes.Grayscale);
 
-        using var features1 = CvDetail.ComputeImageFeatures(featuresFinder, image1);
-        using var features2 = CvDetail.ComputeImageFeatures(featuresFinder, image2);
+        using var features1 = Cv2.Detail.ComputeImageFeatures(featuresFinder, image1);
+        using var features2 = Cv2.Detail.ComputeImageFeatures(featuresFinder, image2);
 
         using var matcher = new AffineBestOf2NearestMatcher();
         using var matchesInfo = matcher.Apply(features1, features2);
