+++ test/OpenCvSharp.Tests/text/DetectTextSWTTest.cs
@@ -11,7 +11,7 @@ public class DetectTextSWTTest : TestBase
         using var src = new Mat("_data/image/imageText.png");
         using var draw = new Mat();
 
-        var rects = CvText.DetectTextSWT(src, true, draw);
+        var rects = Cv2.Text.DetectTextSWT(src, true, draw);
         Assert.NotEmpty(rects);
 
         ShowImagesWhenDebugMode(src, draw);
