+++ test/OpenCvSharp.Tests/xphoto/XPhotoTest.cs
@@ -21,9 +21,9 @@ public class XPhotoTest : TestBase
         using var b = new Mat(src.Rows, src.Cols, src.Type());
         using var g = new Mat(src.Rows, src.Cols, src.Type());
         using var r = new Mat(src.Rows, src.Cols, src.Type());
-        CvXPhoto.ApplyChannelGains(src, b, 2, 1, 1);
-        CvXPhoto.ApplyChannelGains(src, g, 1, 2, 1);
-        CvXPhoto.ApplyChannelGains(src, r, 1, 1, 2);
+        Cv2.XPhoto.ApplyChannelGains(src, b, 2, 1, 1);
+        Cv2.XPhoto.ApplyChannelGains(src, g, 1, 2, 1);
+        Cv2.XPhoto.ApplyChannelGains(src, r, 1, 1, 2);
 
         if (Debugger.IsAttached)
         {
@@ -43,7 +43,7 @@ public class XPhotoTest : TestBase
     [Fact]
     public void GrayworldWBBalanceWhite()
     {
-        using var wb = CvXPhoto.CreateGrayworldWB();
+        using var wb = Cv2.XPhoto.CreateGrayworldWB();
         using var src = LoadImage("lenna.png");
         using var dst = new Mat(src.Rows, src.Cols, src.Type());
         wb.BalanceWhite(src, dst);
@@ -62,7 +62,7 @@ public class XPhotoTest : TestBase
     [Fact]
     public void GrayworldWBProperties()
     {
-        using var wb = CvXPhoto.CreateGrayworldWB();
+        using var wb = Cv2.XPhoto.CreateGrayworldWB();
         var saturationThreshold = wb.SaturationThreshold;
 
         const float val = 100f;
@@ -78,7 +78,7 @@ public class XPhotoTest : TestBase
         using var src = LoadImage("building.jpg");
         using var mask = LoadImage("building_mask.bmp", ImreadModes.Grayscale);
         using var dst = new Mat(src.Size(), src.Type());
-        CvXPhoto.Inpaint(src, mask, dst, OpenCvSharp.XPhoto.InpaintTypes.SHIFTMAP);
+        Cv2.XPhoto.Inpaint(src, mask, dst, OpenCvSharp.XPhoto.InpaintTypes.SHIFTMAP);
         ShowImagesWhenDebugMode(src);
         ShowImagesWhenDebugMode(dst);
     }
@@ -86,7 +86,7 @@ public class XPhotoTest : TestBase
     [Fact]
     public void LearningBasedWBBalanceWhite()
     {
-        using var wb = CvXPhoto.CreateLearningBasedWB(null);
+        using var wb = Cv2.XPhoto.CreateLearningBasedWB(null);
         using var src = LoadImage("lenna.png");
         using var dst = new Mat(src.Rows, src.Cols, src.Type());
         wb.BalanceWhite(src, dst);
@@ -132,7 +132,7 @@ public class XPhotoTest : TestBase
     {
         // About model file
         // http://docs.opencv.org/trunk/dc/dcb/tutorial_xphoto_training_white_balance.html
-        using var wb = CvXPhoto.CreateLearningBasedWB("");
+        using var wb = Cv2.XPhoto.CreateLearningBasedWB("");
         using var src = LoadImage("lenna.png");
         using var dst = new Mat(src.Rows, src.Cols, src.Type());
         wb.BalanceWhite(src, dst);
@@ -174,7 +174,7 @@ public class XPhotoTest : TestBase
     [Fact]
     public void SimpleWBBalanceWhite()
     {
-        using var wb = CvXPhoto.CreateSimpleWB();
+        using var wb = Cv2.XPhoto.CreateSimpleWB();
         using var src = LoadImage("lenna.png");
         using var dst = new Mat(src.Rows, src.Cols, src.Type());
         wb.BalanceWhite(src, dst);
@@ -225,7 +225,7 @@ public class XPhotoTest : TestBase
     {
         using var src = LoadImage("lenna.png");
         using var dst = new Mat();
-        CvXPhoto.DctDenoising(src, dst, 1);
+        Cv2.XPhoto.DctDenoising(src, dst, 1);
 
         if (Debugger.IsAttached)
         {
@@ -238,7 +238,7 @@ public class XPhotoTest : TestBase
     {
         using var src = LoadImage("lenna.png", ImreadModes.Grayscale);
         using var dst = new Mat();
-        CvXPhoto.Bm3dDenoising(src, dst);
+        Cv2.XPhoto.Bm3dDenoising(src, dst);
 
         if (Debugger.IsAttached)
         {
@@ -251,7 +251,7 @@ public class XPhotoTest : TestBase
     {
         using var src = LoadImage("lenna.png", ImreadModes.Grayscale);
         using var dst = new Mat();
-        CvXPhoto.OilPainting(src, dst, 5, 10);
+        Cv2.XPhoto.OilPainting(src, dst, 5, 10);
 
         if (Debugger.IsAttached)
         {
